### 2025-03-31 #7966 @editorialbot
[PRE REVIEW]: HeXtractor: Extracting Heterogeneous Graphs from Structured and Textual Data for Graph Neural Networks
**Submitting author:** @maddataanalyst (<a href="http://orcid.org/0000-0001-5938-7260">Filip Wójcik</a>)
**Repository:** https://github.com/maddataanalyst/hextractor
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.1
