### 2025-04-14 #7966 @Nikoleta-v3
Hey @maddataanalyst 👋🏻 Thank you very much for the submission to JOSS! As you can see here:
https://github.com/openjournals/joss-reviews/issues/7966#issuecomment-2766848681, several of the references are missing DOIs.

Could you please go over the ones without a DOI and check whether they actually have one? Once you've updated the `.bib` file and pushed the changes, you can comment on the issue with:

Thank you!
