### 2025-04-14 #7966 @maddataanalyst
> Hey [@maddataanalyst](https://github.com/maddataanalyst) 👋🏻 Thank you very much for the submission to JOSS! As you can see here: [#7966 (comment)](https://github.com/openjournals/joss-reviews/issues/7966#issuecomment-2766848681), several of the references are missing DOIs.
>
> Could you please go over the ones without a DOI and check whether they actually have one? Once you've updated the `.bib` file and pushed the changes, you can comment on the issue with:
>
> ```
> @editorialbot check references
> ```
>
> Thank you!

Sure, not a problem at all. I will look for the missing DOIs and let you know. Best regards and thank you for starting the review.
