### 2025-04-14 5d71bac
Updated references according to the suggestions from the editorial bot.
