+++ paper.bib
@@ -1,223 +1,234 @@
 @inbook{Shi2022,
-   abstract = {Heterogeneous graphs (HGs) also called heterogeneous information networks (HINs) have become ubiquitous in real-world scenarios. Recently, employing graph neural networks (GNNs) to heterogeneous graphs, known as heterogeneous graph neural networks (HGNNs) which aim to learn embedding in low-dimensional space while preserving heterogeneous structure and semantic for downstream tasks, has drawn considerable attention. This chapter will first give a brief review of the recent development on HG embedding, then introduce typical methods from the perspective of shallow and deep models, especially HGNNs. Finally, it will point out future research directions for HGNNs.},
-   author = {Chuan Shi},
-   doi = {10.1007/978-981-16-6054-2_16},
-   editor = {Lingfei Wu and Peng Cui and Jian Pei and Liang Zhao},
-   isbn = {9789811660542},
-   booktitle = {Graph Neural Networks: Foundations, Frontiers, and Applications},
-   month = {1},
-   pages = {351-369},
-   publisher = {Springer Nature},
-   title = {Heterogeneous Graph Neural Networks},
-   year = {2022}
+	title        = {Heterogeneous Graph Neural Networks},
+	author       = {Chuan Shi},
+	year         = 2022,
+	month        = 1,
+	booktitle    = {Graph Neural Networks: Foundations, Frontiers, and Applications},
+	publisher    = {Springer Nature},
+	pages        = {351--369},
+	doi          = {10.1007/978-981-16-6054-2_16},
+	isbn         = 9789811660542,
+	abstract     = {Heterogeneous graphs (HGs) also called heterogeneous information networks (HINs) have become ubiquitous in real-world scenarios. Recently, employing graph neural networks (GNNs) to heterogeneous graphs, known as heterogeneous graph neural networks (HGNNs) which aim to learn embedding in low-dimensional space while preserving heterogeneous structure and semantic for downstream tasks, has drawn considerable attention. This chapter will first give a brief review of the recent development on HG embedding, then introduce typical methods from the perspective of shallow and deep models, especially HGNNs. Finally, it will point out future research directions for HGNNs.},
+	editor       = {Lingfei Wu and Peng Cui and Jian Pei and Liang Zhao}
 }
 @inproceedings{Hu2020,
-   author = {Ziniu Hu and Yuxiao Dong and Kuansan Wang and Yizhou Sun},
-   booktitle = {Proceedings of the web conference 2020},
-   pages = {2704-2710},
-   title = {Heterogeneous graph transformer},
-   year = {2020}
+	title        = {Heterogeneous graph transformer},
+	author       = {Ziniu Hu and Yuxiao Dong and Kuansan Wang and Yizhou Sun},
+	year         = 2020,
+	booktitle    = {Proceedings of the web conference 2020},
+	pages        = {2704--2710},
+	doi          = {10.1145/3366423.3380027}
 }
 @article{Yang2020,
[…]
