### 2025-04-14 #7966 @maddataanalyst
Looks good now. For this the book by Mayank Kejriwal, Craig A. Knoblock, Pedro Szekely, "Knowledge Graphs: Fundamentals, Techniques, and Applications" MIT Press, 2021 - I cannot find the DOI anywhere, probably it was not assigned.
