### 2025-05-02 #8057 @jboynyc
Regarding the installation instructions: Why assume that everyone is using Anaconda as their Python distribution? Requiring `conda` *and* `poetry` feels redundant (and I was able to get things up and running without using `conda`).
