### 2025-05-02 maddataanalyst/hextractor#13 @jboynyc
installation instructions
As part of openjournals/joss-reviews#8057 I took at look at the documentation and I had a question about the installation instructions. Why assume that everyone is using Anaconda as their Python distribution? Requiring conda and poetry feels redundant (and I was able to get things up and running without using conda). I also see that your GitHub workflow does not require conda either.
