### 2025-05-02 maddataanalyst/hextractor#14 @jboynyc
contributor instructions
As part of openjournals/joss-reviews#8057: JOSS wants your project to include clear community guidelines, i.e., clear guidelines for third parties wishing to

1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

Consider including a `CONTRIBUTING` file in the top level of your repository. You may consider adapting [this template](https://github.com/audreyfeldroy/cookiecutter-pypackage/blob/master/CONTRIBUTING.rst).
