### 2025-05-06 66d2e04
Fixed installation instructions after the JOSS review.
