### 2025-05-07 8e6ea7c
Added the CONTRIBUTING.rst file according to the review, and updated the docs.
