+++ docs/contributing.md
@@ -1,0 +1,173 @@
+# Contributing and help
+
+Contributions are welcome, and they are greatly appreciated! Every little bit helps, and credit will always be given.
+
+You can contribute in many ways:
+
+## Types of Contributions
+
+### Report Bugs
+
+Report bugs at https://github.com/maddataanalyst/hextractor/issues
+
+If you are reporting a bug, please include:
+
+* Your operating system name and version.
+* Any details about your local setup that might be helpful in troubleshooting.
+* Detailed steps to reproduce the bug.
+
+### Fix Bugs
+
+Look through the GitHub issues for bugs. Anything tagged with "bug" and "help wanted" is open to whoever wants to implement a fix for it.
+
+### Implement Features
+
+Look through the GitHub issues for features. Anything tagged with "enhancement" and "help wanted" is open to whoever wants to implement it.
+
+### Write Documentation
+
+HeXtractor could always use more documentation, whether as part of the official docs, in docstrings, or even on the web in blog posts, articles, and such.
+
+### Submit Feedback
+
+The best way to send feedback is to file an issue at https://github.com/maddataanalyst/hextractor/issues.
+
+If you are proposing a new feature:
+
+* Explain in detail how it would work.
+* Keep the scope as narrow as possible, to make it easier to implement.
+* Remember that this is a volunteer-driven project, and that contributions are welcome :)
+
+## Get Started!
+
+Ready to contribute? Here's how to set up `hextractor` for local development. Please note this documentation assumes you already have `Git` installed and ready to go.
+
+1. Fork the `hextractor` repo on GitHub.
+
+2. Clone your fork locally:
+    ```bash
+    cd path_for_the_repo
+    git clone git@github.com:YOUR_NAME/hextractor.git
+    ```
+
+3. Set up your development environment. You can choose between Conda or a standard Python virtual environment:
+
+    **Option 1: Using Conda**
+    ```bash
+    # Create a new conda environment from the provided file
+    conda env create -f environment.yml
+    
+    # Activate the environment
+    conda activate hextractor
+    
+    # Install poetry inside the conda environment
+    pip install poetry
+    
+    # Install the package with all dependencies
+    poetry install --with dev --with research
+    ```
+
+    **Option 2: Using Standard Python Virtual Environment**
+    ```bash
+    # Using venv (Python 3.3+)
+    python -m venv hextractor-env
+    source hextractor-env/bin/activate  # On Windows: hextractor-env\Scripts\activate
+    
+    # Or using virtualenv
+    virtualenv hextractor-env
+    source hextractor-env/bin/activate  # On Windows: hextractor-env\Scripts\activate
+    
+    # Install poetry
+    pip install poetry
+    
+    # Install the package with all dependencies
+    poetry install --with dev --with research
+    ```
+
+    This should change the shell to look something like:
+    ```bash
+    (hextractor) $  # or (hextractor-env) $
+    ```
+
+4. Create a branch for local development:
+    ```bash
+    git checkout -b name-of-your-bugfix-or-feature
+    ```
+
+    Now you can make your changes locally.
+
+5. When you're done making changes, check that your changes pass the tests:
+    ```bash
+    pytest ./tests
+    ```
+
+6. The next step would be to run all test cases. Before you run pytest, ensure all dependencies are installed:
+    ```bash
+    # Dependencies should already be installed if you used poetry install as instructed above
+    pytest ./tests
+    ```
+
+    If you get any errors while installing packages, try updating pip:
+    ```bash
+    # Update pip
+    pip install -U pip
+    ```
+
+7. Before raising a pull request you should also run tox. This will run the tests across different versions of Python:
+    ```bash
+    tox
+    ```
+
[…]
