### 2025-05-09 maddataanalyst/hextractor#13 @maddataanalyst
@jboynyc , we have changed the installation instructions, according to your remarks. Thank you!
