### 2025-05-09 maddataanalyst/hextractor#14 @maddataanalyst
@jboynyc , we have added the CONTRUBUTING instructions, according to your remarks. We have also added similar chapter to the documentation. Thank you!
