### 2025-05-09 #8057 @maddataanalyst
> > Regarding the installation instructions: Why assume that everyone is using Anaconda as their Python distribution? Requiring `conda` _and_ `poetry` feels redundant (and I was able to get things up and running without using `conda`).
>
> [@jboyny](https://github.com/jboyny), hello, and thank you for your feedback. I completely agree— from a technical standpoint, there is no need to duplicate functionality between Conda and Poetry, nor should we assume that all users rely on this combination. While many data scientists do use it (and most of the commercial projects we’re involved in adopt this setup: Conda environments alongside Poetry), it’s not a requirement for an open-source library.
>
> We will update the README and documentation to reflect this. Thank you again for the valuable suggestion.
>
> We’re currently enjoying an extended bank holiday in our country (with great weather, fortunately!), so [@mmalczewski](https://github.com/mmalczewski) and I will address these changes starting Monday, May 6.
>
> Best regards!

We have resolved both issues, changing three places:
1. README file;
2. Installation instructions;
3. Documentation.

Thank you for your remarks!
