### 2025-06-04 #8057 @cjbarrie
Reviewing now. Really nice software library.

Some notes:

1. As @jboynyc, the README is confusing as it advises users to install with poetry. But unless the user has the source code locally, they're going to get the error:

```
Poetry could not find a pyproject.toml file in /Users/christopherbarrie/Dropbox/sandbox/hextractor or its parents
```

So if the README is just telling users how to install the library, why advise using Poetry?

2. The library is using Pydantic v1 syntax so the user is often exposed to this type of warning:

```
hextractor-env(base) christopherbarrie@SOC-CB5691-MBP16-24 hextractor % python test.py
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:30: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  company_node_params = structures.NodeTypeParams(
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:38: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  company_tags_node_params = structures.NodeTypeParams(
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:45: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  employee_node_params = structures.NodeTypeParams(
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:54: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  company_has_emp_edge_params = structures.EdgeTypeParams(
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:62: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  company_has_tag_edge_params = structures.EdgeTypeParams(
```

3. There are no community guidelines in the README
