### 2025-06-05 #8057 @maddataanalyst
Hello!
Thank you for the review and kind words :)

> 1. As [@jboynyc](https://github.com/jboynyc), the README is confusing as it advises users to install with poetry. But unless the user has the source code locally, they're going to get the error:
>
> ```
> Poetry could not find a pyproject.toml file in /Users/christopherbarrie/Dropbox/sandbox/hextractor or its parents
> ```
>
> So if the README is just telling users how to install the library, why advise using Poetry?

Ahhh you're right. When writing this I was looking from the perspective of the **library developer** not **the user**. We will rewrite the README and clearly state two types of installation - simple using `pip install` and from the source code directly.

> 2. The library is using Pydantic v1 syntax so the user is often exposed to this type of warning:
>
> ```
> hextractor-env(base) christopherbarrie@SOC-CB5691-MBP16-24 hextractor % python test.py
> /Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:30: UserWarning: A custom validator is returning a value other than `self`.
> Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
> See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
>   company_node_params = structures.NodeTypeParams(

Thank you for spotting this warning. We will fix it + update Pydantic in the project as well.

> 3. There are no community guidelines in the README

Sure we will add this.

I will create and link proper issues in our repo + let you know once all is done.
