### 2025-06-05 maddataanalyst/hextractor#17 @maddataanalyst
Add community guidelines to README
As pointed out in the review openjournals/joss-reviews#8057, community guidelines should be added to the README.
