### 2025-06-05 maddataanalyst/hextractor#18 @maddataanalyst
Remove Pydantic warnings
As pointed out in the review openjournals/joss-reviews#8057, some modules that use validators, throw a Pydantic warning message. Custom validators need to be updated.

Example from the review

```
hextractor-env(base) christopherbarrie@SOC-CB5691-MBP16-24 hextractor % python test.py
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:30: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  company_node_params = structures.NodeTypeParams(
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:38: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  company_tags_node_params = structures.NodeTypeParams(
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:45: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  employee_node_params = structures.NodeTypeParams(
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:54: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  company_has_emp_edge_params = structures.EdgeTypeParams(
/Users/christopherbarrie/Dropbox/sandbox/hextractor/test.py:62: UserWarning: A custom validator is returning a value other than `self`.
Returning anything other than `self` from a top level model validator isn't supported when validating via `__init__`.
See the `model_validator` docs (https://docs.pydantic.dev/latest/concepts/validators/#model-validators) for more details.
  company_has_tag_edge_params = structures.EdgeTypeParams(

```
