### 2025-06-05 maddataanalyst/hextractor#19 @maddataanalyst
Improve installation instructions
As pointed out in the review openjournals/joss-reviews#8057, installation instructions need to me more precise.

Probably two separate sub-sections are needed: installation from PyPI and installation from source using Pydantic.

Direct review quote:

> As @jboynyc, the README is confusing as it advises users to install with poetry. But unless the user has the source code locally, they're going to get the error:
> Poetry could not find a pyproject.toml file in /Users/christopherbarrie/Dropbox/sandbox/hextractor or its parents
