### 2025-06-05 11fef44
Added community guidelines to the README file
