### 2025-06-05 4d75212
Added direct reference to the Contributing.rst file in README.
