### 2025-06-05 #8057 @maddataanalyst
@cjbarrie , we have updated the project and added 3 changes:
1. Annoying warning from pydantic validators is gone.
2. README file has changed:
   1. Explicitly mentioned two installation types: recommended - from pypi via `pip install` and a separate mode - for developers and contributor - from the source code.
   2.  As the file `CONTRIBUTING.rst` is already in the repository - it was explicitly mentioned + linked in the README.md, as the source of the community guidelines and other information for contributors.
