### 2025-06-10 58884de
Corrected typos and other quirks in the documentation.
I noticed these in the course of the JOSS review process at openjournals/joss-reviews#8057
