+++ docs/reference/data_sources.md
@@ -1,5 +1,3 @@
 # Data Sources
 
 ::: hextractor.data_sources
-    options:
-        show_symbol_type_toc: true
