### 2025-06-18 f07c1cd
Changed the date to today in paper.
