+++ paper.md
@@ -19,7 +19,7 @@ affiliations:
    index: 1
  - name: Diveapps, Wrocław, Poland
    index: 2
-date: 24 March 2025
+date: 18 June 2025
 bibliography: paper.bib
 ---
 
