### 2025-06-18 #8057 @maddataanalyst
@Nikoleta-v3 thank you for the information and assistance. Below you can find all the details:

1. I have updated the paper with an additional dot :) Also triggered the generation of the new pdf.
2. We have released a new software version along with the review updates:
    - Old version (the original submission): v1.0.1
    - **New version (with all corrections): v1.0.2**
3. Software is uploaded to the Zenodo:
    - Upload: https://zenodo.org/records/15687595
    - DOI: 10.5281/zenodo.15687595

One more question from my side + one information:

1. Question: once we are done with the paper review and acceptance in JOSS - can we delete the "paper" branch and move all its content to the main branch? If not it's not a problem of course :)
2. I will be on holidays starting today afternoon (18.06.2025) up to the end of the week (22.06.2025) so I'll respond to any remarks next Monday.

Thank you again for help and assistance!
