### 2025-06-18 #8057 @Nikoleta-v3
Thank you for including the `dot`! You can delete the paper branch if you'd like 😄

Please enjoy your holidays! 🏖  I'm just going to run a few final commands and do one last round of checks. If anything comes up, I’ll leave a comment and you can get back to me on Monday.

If not (and probably not!), I’ll go ahead and recommend the paper for acceptance. After that, the handling editor will give it one final look before publishing.
