### 2025-06-19 5914aad
remove duplicate entry from bib
