+++ paper.bib
@@ -74,15 +74,6 @@
 	isbn         = 9783031019180,
 	url          = {https://books.google.pl/books?id=4YdyEAAAQBAJ}
 }
-@article{Shi2022,
-	title        = {Heterogeneous graph neural networks},
-	author       = {Chuan Shi},
-	year         = 2022,
-	journal      = {Graph Neural Networks: Foundations, Frontiers, and Applications},
-	publisher    = {Springer},
-	pages        = {351--369},
-	doi          = {10.1007/978-981-16-6054-2_16}
-}
 @book{Kejriwal2021,
 	title        = {Knowledge graphs: Fundamentals, techniques, and applications},
 	author       = {Mayank Kejriwal and Craig A Knoblock and Pedro Szekely},
