### 2025-06-19 #8057 @Nikoleta-v3
`and probably not!` famous last words 😄 😆

Unfortunately, there was an issue when I tried to recommend acceptance. It turns out there's a duplicate entry in your `.bib` file. I’ve opened a pull request to remove it: https://github.com/maddataanalyst/hextractor/pull/24.

Please go ahead and accept the PR once you're back, if it looks good to you. Then I can try to start the acceptance process again 🙌🏻  Thank you!
