### 2025-06-23 #8057 @maddataanalyst
Hahaha, not a problem at all :D Thank you for fixing this issue on the spot - `.bib` files are terrible to work with IMHO - always some duplicates or malformed entries :D
I have merged your pull request, hopefully we can proceed now.

Best regards!
